requirements.txt
# Core dependencies
holoocean>=0.5.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
pyyaml>=5.4.0
scipy>=1.7.0

# Data processing
pandas>=1.3.0
scikit-image>=0.18.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Metrics
scikit-learn>=0.24.0

# Optional: Experiment tracking
# mlflow>=1.20.0
# wandb>=0.12.0
